MITx-6.00.1x-Introduction-to-Computer-Science-and-Programming-Using-Python
Welcome to 6.00.1x! In this course you’ll be learning the basics of
computer programming in Python and the fundamentals of computation, as
well as getting the opportunity to implement your own Python functions.

This course is offered online and we understand that there are many
opportunities available to cheat. We caution you to not do so. You will
learn less and only harm yourself by cheating. We ask that you review
our collaboration and forum guidelines, available on the course
handouts page, to understand how we expect our students to conduct
themselves in this course. Additionally all students are expected to
follow the edX Honor Code, available at https://www.edx.org/honor